MemberController 불필요한 로깅 제거

+++ practice/chaeeun/demo/src/main/java/com/example/demo/service/MemberService.java
@@ -3,6 +3,7 @@ package com.example.demo.service;
 import com.example.demo.domain.Member;
 import com.example.demo.dto.MemberDTO;
 import com.example.demo.exception.DuplicateEntityException;
+import com.example.demo.exception.NotFoundException;
 import com.example.demo.repository.MemberRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -21,16 +22,17 @@ public class MemberService {
         this.memberRepository = memberRepository;
     }
 
+    private void validateDuplicateEmail(String email) {
+        if(memberRepository.existsByEmail(email)) {
+            throw new DuplicateEntityException("이미 등록된 회원입니다.");
+        }
+    }
     @Transactional
     public MemberDTO registerMember(MemberDTO memberDTO) {
-        if(memberRepository.existsByEmail(memberDTO.getEmail())) {
-            log.error("이미 등록된 회원입니다. email={}", memberDTO.getEmail());
-            throw new DuplicateEntityException("이미 등록된 회원입니다.");
-
-        }
+        validateDuplicateEmail(memberDTO.getEmail());
         Member member = new Member(memberDTO.getName(), memberDTO.getEmail());
         Member savedMember = memberRepository.save(member);
-        log.info("회원 등록에 성공하였습니다. id:{}, name:{}, email:{}", savedMember.getId(), savedMember.getName(), memberDTO.getEmail());
+        log.info("새 회원 등록: name={}, email={}", savedMember.getName(), savedMember.getEmail());
         return new MemberDTO(savedMember.getId(), savedMember.getName(), savedMember.getEmail());
     }
 
@@ -45,7 +47,9 @@ public class MemberService {
     // 회원 삭제
     @Transactional
     public void deleteMember(Long id) {
-        log.info("회원 삭제에 성공하였습니다. id={}", id);
+        if(!memberRepository.existsById(id)) {
+            throw new NotFoundException("존재하지 않는 회원입니다. id=" + id);
+        }
         memberRepository.deleteById(id);
     }
